+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres (more or less) to [Semantic Versioning](http://semver.o
 When you submit a PR, add your changes here!
 
 ### Added
-- contentComponent prop to allow specifying a custom component to render the items
+- itemRenderer prop to allow specifying a custom component to render the items
 
 ## [0.11.1]
 ### Fixed